+++ src/components/data/DataTable.js
@@ -13,7 +13,6 @@ import {
   Link,
   RadioButton,
 } from 'src/components/common';
-import { concatenateAttributeNames } from 'src/components/data/attribute-utils';
 import { EditDataLink, entityAttributeText, EntityRenamer, HeaderOptions, renderDataCell, SingleEntityEditor } from 'src/components/data/data-utils';
 import RenameColumnModal from 'src/components/data/RenameColumnModal';
 import {
@@ -28,6 +27,7 @@ import { MenuButton } from 'src/components/MenuButton';
 import Modal from 'src/components/Modal';
 import { MenuTrigger } from 'src/components/PopupTrigger';
 import { GridTable, HeaderCell, paginator, Resizable, TooltipCell } from 'src/components/table';
+import { concatenateAttributeNames } from 'src/data/data-table/entity-service/attribute-utils';
 import { Ajax } from 'src/libs/ajax';
 import colors from 'src/libs/colors';
 import { withErrorReporting } from 'src/libs/error';
