+++ src/components/data/data-utils.js
@@ -19,7 +19,6 @@ import {
   Select,
   spinnerOverlay,
 } from 'src/components/common';
-import { convertAttributeValue, getAttributeType } from 'src/components/data/attribute-utils';
 import Dropzone from 'src/components/Dropzone';
 import { icon } from 'src/components/icons';
 import { AutocompleteTextInput, ConfirmedSearchInput, PasteOnlyInput, TextInput, ValidatedInput } from 'src/components/input';
@@ -31,6 +30,7 @@ import { Sortable, TextCell } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
 import { isAzureUri, isDrsUri, isGsUri } from 'src/components/UriViewer/uri-viewer-utils';
 import { UriViewerLink } from 'src/components/UriViewer/UriViewerLink';
+import { convertAttributeValue, getAttributeType } from 'src/data/data-table/entity-service/attribute-utils';
 import AttributeInput, { AttributeTypeInput } from 'src/data/data-table/entity-service/AttributeInput';
 import ReferenceData from 'src/data/reference-data';
 import { Ajax } from 'src/libs/ajax';
